Sometimes you just commit whatever you did
It doesn't really work at all.  Or build.
I decided to start supporting sqlite because it seemed easier
to debug with a local copy of that.  Then I decided to try to
make a unified interface but because I don't really know how
the "interface" functionality in Go works I just decided to use
function pointers like a real C programmer.